+++ examples/readme_example.rs
@@ -1,12 +1,12 @@
 use std::cell::RefCell;
 use std::rc::Rc;
 use toy_ui::{
-    app,
+    app::app,
     color::colors,
-    draw::TextStyle,
     element::{container, row, text},
     interaction::Interactable,
     layer::{LayerOptions, MouseButton},
+    style::TextStyle,
 };
 
 fn main() {
