+++ src/interaction/mod.rs
@@ -1,7 +1,9 @@
 //! Interaction system for handling mouse events with z-order based hit testing
 
-use crate::geometry::Point;
-use crate::layer::{InputEvent, MouseButton};
+use crate::{
+    geometry::Point,
+    layer::{InputEvent, MouseButton},
+};
 use glam::Vec2;
 use std::collections::HashMap;
