+++ src/main.rs
@@ -1,9 +1,9 @@
 use palette::Srgba;
 use toy_ui::{
-    app,
-    draw::TextStyle,
+    app::app,
     element::{column, container, row, text},
     layer::{LayerManager, LayerOptions},
+    style::TextStyle,
 };
 use tracing::{info, info_span};
 use tracing_subscriber::{EnvFilter, fmt};
